services/supply_engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
firebase-admin==6.4.0
python-dotenv==1.0.0
pandas==2.2.0
beautifulsoup4==4.12.3
lxml==5.1.0
Pillow==10.2.0